+++ database/migrations/2025_04_13_071306_create_incomes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('incomes', function (Blueprint $table) {
             $table->id();
-            $table->double('total');
+            $table->integer('total');
             $table->timestamps();
         });
     }
